Move teams form to separate component, correct layers manager route
