+++ src/components/layers-manager/LayersForm.js
