+++ src/components/layers-manager/LayersManager.js
