+++ src/components/layers-manager/LayersSwitcher.js
