+++ src/components/layers-manager/LayersSwitcherContainer.js
